Refactor show logic into repository
Moves some of the code into the repository. The show command was not the
best place for this. Writing tests required that we move some default
loaded tilbits into separate arguments.

Since now a lot of those functions have the same second argument, that
suggest to me that the repository might be a type holding the tilbits
and those functions should act on it. Leaving this for a later refactor
after specs are written for most commands.