Remove the List command
The list and show commands started to converge quite a while ago.
Currently the list command was just a debug tool with a different output
format.

Since the show command was more robust, all that was needed was to add
the yaml format to the command.

There were considerable issues with binding the formatting flag to both
the root and show commands, since the show command overrode the default
values. Perhaps a persistent flag approach for the root command should
be used in this case.