Move current methods into a common LocalSourcesRepository
The methods that were moved were working on a passed in array of Tilbit
objects. This was a smell that they should be working on a struct that
has these arrays already loaded. It simplified the code and the tests.

It also makes it possible to start working on an implementation of
another repository object for SQLite which will have the same Repository
interface and be able to switch them pretty easily, compare their
results etc.

Ultimately, the LocalSourcesRepository type will be removed from the
source code.

As part of the refactoring, the `sources` command was removed since it
was mostly used for debugging purposes and is no longer useful.