Implement and test SQLIte repository
This needed quite a few changes to the surrounding interfaces, but was
finally done. Some of the changes made seem symptomatic of the old way
of doing things, but we will handle those as we go along.

In the meantime, the specs for the repos are DB by default, but other
parts of the system still use the old file based repo. This is fine for
now, but the old system will probably be removed in the next changes,
with the changes needed for it's removal to be done first.