Supply an option to provide a specific directory.
This is another in the epic of adding flexibility to both the repos and
the parsers. I'm currently going back and forth with the tradeoffs that
adding a database layer will bring.

I like the idea of having the flexibility of adding additional features
into the tool (tags, notes, reliable recurrence etc), but I'm hesitant
to give up on the simplicity of just passing in a directory with flat
files.

So I'm currently keeping my options open and making the architecture
open enought to provide several sourcing options. I'm currently leaning
towards moving away from the simple object based repo and moving the
functinality into source parsers, the repo will be left as a layer for
finding specific objects.