Make the move to sqlite repos
This commit does several things:
- Makes the repos lighter by removing the parsing logic entitely.
- Parsers are now handling the parsing of sources.
- The only repository left is the sqlite based one.

This opens up a future where there are several sources (input), but also
a place to have ones notes and additional data.

At some point, it's easier to have stuff in a database then to deal with
flat files. Ideally, the actual tilbits can be shared between people,
but they all can have their own notes for each one. This system allows
this.

I plan to have an arbitary list of input sources that are read-only and
one sqlite-based write target.