Render markdown content properly using glamour
Another library that takes the load off. The code is pretty gnarly for
now, but let's not dwell on this.

Next steps will probably be some kind of nicer outputs, especially
regarding source data. Probably make the source into some kind of title
bar, to save on space inside the box itself.