Add colorful formatting option
This jives more with what I want the tool to look like. Ultimately, this
will be the default option. There's a lot of options from this point
forward, like customizing the colors and the elements in the title bars.

A possible option would be to add a bottom bar for some additional info.